GeneSetReader: Fix parsing
Call trim on the read line, use boost::token_compress_on to allow multiple
whitespaces as separator and fix the loop counter. Do not call trim in
Processors, as the input is now guaranteed to be properly trimmed.

Also only allocate the result vector once.

Oh, and fix a grammar mistake..