WeightedGSEA: speed up PermuationTest significantly
The idea is to shuffle exactly once for every permutation.
We shuffle the indices using the largest sampleSize. We assume the indices
to be sorted in the range [0, sampleSize).
Using this invariant we just need to sort the range
[lastSampleSize, currentSampleSize) and can simply merge the two intervals.

This gives us a factor of about three for large permutation counts.