[core] Fix compiler warning in DependentTTest
The end of the second input range was never validated. We now check
that both ranges contain the same number of elements and throw if not.

This also means that the code is slightly less efficient for linked
list inputs.

We then use std::transform to initialize the difference vector.