Complete rework of enrichment computation
This lands the groundwork for enabling ColumnWise p-Values etc. for all
Enrichment methods.

The basic idea is to introduce SetLevelStatistics adapter classes, that
unify the interfaces of the enrichment algorithms. These are then wrapped
into an instance of the EnrichmentAlgorithm class, which is automatically
generated using templates.

Unfortunately this may lead to some regressions concerning computation
speed as well as breakage for some algorithms.