[core] Update Category to take a raw EntityDatabase pointer
This may change again in the future, but the current idea is, that
Categories are worthless without an associated EntityDatabase. Due
to this, the user needs to keep an instance of EntityDatabase around
from the start of the program. Having a lot of heavy-weight shared_ptrs
around only consumes memory and is detrimental to performance.