Implement the GSEA p-value computation method.
This first normalizes all enrichment scores and then directly
computes an FDR q-value for every test.

This commit also fixes a long standing bug with sorting the
scores in the proper order. We previously unconditionally sorted
the scores in ascending order, which led to subtly skewed p-values
in the case of weighted GSEA.