[core] Scores::subsetMerge: fix criterion checking that an entry was found
We simply need to check that the iterator has not reached the end. The
previous check could lead to dereferencing an invalid pointer. This bug
had the potential to be catastrophic, however in practice not much should
have gone wrong as the memory pointed to by the iterator was with _very_
high probability had a value that was different from the one we searched.